More bugfixes and better handling of MySQL errors.
git-svn-id: https://halldsvn.jlab.org/repos/trunk/scripts/monitoring@16574 e333a0bb-b310-0410-bd06-bbc7ae8e47d2

+++ process/datamon_db.py
@@ -53,6 +53,8 @@ class datamon_db:
         ## be careful with this!!
         #self.db.execute("DROP TABLE *")  
         for table in self.table_names:
+            if table == "run_info" or table == "version_info":
+                continue
             print "clearing " + table + "..."
             self.db.execute('DELETE FROM ' + table)
         ## reset autoincrement counters
